+++ rector.php
@@ -30,6 +30,7 @@ return RectorConfig::configure()
         __DIR__ . '/tests/Integration/Services/IM',
         __DIR__ . '/src/Services/IMOpenLines',
         __DIR__ . '/tests/Integration/Services/IMOpenLines',
+        __DIR__ . '/tests/Integration/Services/CRM',
         __DIR__ . '/src/Services/Main',
         __DIR__ . '/tests/Integration/Services/Main',
         __DIR__ . '/src/Services/Placement',