+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Added
 
+- Added service `Services\CRM\Automation\Service\Trigger` with support methods,
+  see [add crm.automation.trigger* methods](https://github.com/bitrix24/b24phpsdk/issues/148):
+    - `add` add new trigger
+    - `delete` delete trigger
+    - `list`  get list of triggers
+    - `execute` execute trigger
 - Added **PHP 8.4** [support](https://github.com/bitrix24/b24phpsdk/issues/120) 🚀
 - Added method `Bitrix24\SDK\Services\Main\Service::guardValidateCurrentAuthToken` for validate current auth token with
   api-call `app.info` on vendor OAUTH server.